Crate gloo_console

The JavaScript’s console object provides access to the browser’s console. Using the console object in Rust/WASM directly is cumbersome as it requires JavaScript glue code. This crate exists to solve this problem by providing a set of ergonomic Rust APIs to deal with the browser console.

Example

The following example logs text to the console using console.log

use gloo_console::log;

let object = JsValue::from("any JsValue can be logged");
log!("text", object)

Macros

assert

Calls console.assert()

clear

Calls console.clear()

console

A macro similar to dbg! that logs JsValues to console.

console_dbg

A macro similar to dbg! to log to browser console.

debug

Calls console.debug()

dir

Calls console.dir()

dirxml

Calls console.dirxml()

error

Calls console.error()

group

Calls console.group()

group_end

Calls console.groupEnd()

info

Calls console.info()

log

Calls console.log()

table

Calls console.table()

trace

Calls console.trace()

warn

Calls console.warn()

Structs

Counter

A console time measurement.

Timer

A console time measurement.